refactor: split quantile mapping form
